Split Des_View class into Des_View, Menu_View and Window_View (parent)
