+++ controller/DES/login_button.py
@@ -5,7 +5,7 @@ import sys
 sys.dont_write_bytecode = True
 import PySimpleGUI as sg
 import model.data.accounts as accounts
-import view.data_explorer_view as view
+import view.menu_view as view
 
 def accept(event, values, instance):
     keep_going = True
@@ -13,12 +13,12 @@ def accept(event, values, instance):
         accounts_exists = accounts.verify_account(values['-EMAIL-'], values['-PASSWORD-'])
         if accounts_exists:
             # Add current user
-            view.DES_View.user = accounts.get_display_name(values['-EMAIL-'])
+            view.Menu_View.user = accounts.get_display_name(values['-EMAIL-'])
             # Switch screen
             instance.window['-COL_LOGIN-'].update(visible=False)
             instance.window['-COL_HOME-'].update(visible=True)
             # Update name on home screen
-            view.DES_View.load_home_page(instance)
+            view.Menu_View.load_home_page(instance)
         else:
             sg.Popup('Login failed. Incorrect email or password.')
     return keep_going
