+++ data_explorer.py
@@ -8,7 +8,8 @@
 import sys
 sys.dont_write_bytecode = True # Don't write bytecode files to disk (pyc)
 
-from view.data_explorer_view import DES_View
+from view.window_view import Window_View
+from view.menu_view import Menu_View
 
 
 # Main code
@@ -16,8 +17,8 @@ if __name__ == "__main__" :
     """
         Code that runs when this is the main module.
     """
-    DES_View.set_theme()
-    des_obj = DES_View()
-    des_obj.set_up_layout('main')
+    Window_View.set_theme()
+    des_obj = Menu_View()
+    des_obj.set_up_layout()
     des_obj.render()
-    DES_View.accept_input()
+    Menu_View.accept_input()
